Bugfixes, tested regression on different features
There were a couple of problems with matrix lengths and shuffling on the
data merge script and the linear regression script.

I tried using different sets of discrete features along with the score to
train the linear regression model, which resulted in slightly better
results, and totally awful results when using the publisher.

I don't understand the problem with using the publisher at all.